Refactor build script, debug config and various kernel handling parts
Modified the build-iso.sh script for better debugging by replacing ninja command with ls to merely list files in current directory. Also, adjusted the Global Descriptor Table (GDT) entries in gdt.asm to tweak user mode privilege levels and type of segments used. Debug Kernel run configuration is updated to add the -d int argument to enable interrupt debugging. Adjusted return type of strlen function in misc.c from 'int' to 'unsigned long int' for better compatibility with length manipulation. Additional testing code has been added to kernel.c and a no-op for free function in the bump allocator in alloc.c has been implemented to prevent potential memory leak issues.

+++ klibc/include/string.h
@@ -303,7 +303,7 @@ char *strncat(char *dest, const char *src, size_t n);
  *
  * @return This function returns the length of string.
  */
-int	 strlen(const char *s);
+long unsigned int	 strlen(const char *s);
 
 /**
  * @brief Converts the input string to its corresponding integer value.
